requirements.txt
sentencepiece
accelerate
transformers
diffusers
openai-clip
qwen_vl_utils
matplotlib